refactor(settings): extract the 37 /settings routes into src/settings/routes.py
Pure move, no behaviour change. The 37 route-methods whose path starts with
/settings were the largest block of handlers still defined inline in
src/api_server.py with no router tag. They now live in
src/settings/routes.py behind APIRouter(tags=["settings"]), mounted with
app.include_router alongside pages/schedules/collections.

Moved verbatim: the 37 handlers plus the four definitions that exist only to
serve them — SilenceScheduleRequest, BoardIdentifyRequest,
_beta_https_status, _hdmi_kiosk_supported.

The only edits inside a moved body are the ones relocation forces:
  - @app.<verb> -> @router.<verb>
  - relative imports re-anchored (from .devices -> from src.devices, and
    from .settings.service -> from .service)
  - two api_server module globals read back at call time:
    _service_running and PLUGIN_SYSTEM_AVAILABLE

Verified mechanically, not by eye: all 41 moved definitions are AST-identical
(ast.unparse) to their originals at HEAD~ after normalising exactly those
four edits.

Collaborators stay on the api_server seam, deliberately. /settings reaches 25
collaborators that either live only in api_server (get_service,
_require_board, _reinitialize_board_clients, the updater-sidecar probes) or
that other, unconverted routers still resolve through api_server —
get_settings_service alone is patched at 160 sites, get_service at 95,
PLUGIN_SYSTEM_AVAILABLE at 88. Per API_CONVENTIONS.md ("during extraction,
moved handlers resolve api_server-patched names at call time") and the Phase 2
plan's addendum 3 ("a shared accessor cannot be deleted until its last
consumer converts"), those resolve through a one-line _api() seam. Only
collaborators with a canonical home that nothing patches via api_server
(classify_dimensions, run_board_send, VALID_STRATEGIES,
VALID_OUTPUT_TARGETS) are imported normally.

src/ops/executors.py called six of these handlers off `src.api_server`
directly; it now imports them from src.settings.routes.

Gate (this commit is the safety gate for the conversion commits that follow):
  - tests/golden/api_routes.json byte-identical (route inventory records
    path/methods/name/kind, and none of the three changed)
  - full suite before: 1 failed, 5886 passed, 2 skipped
    full suite after:  1 failed, 5886 passed, 2 skipped
    (the one failure is the known worktree-environmental
     test_check_image_formats.py::TestRepositoryIsClean)
  - data-dir leak count: 0
  - ruff 0.16.5 check + format clean

Refs: Phase 2 Task 8 (.superpowers/plans/2026-09-06-rework-phase2.md)
