+++ pyproject.toml
@@ -174,6 +174,7 @@ ignore = [
 "src/system/update_service.py" = ["PTH", "SIM", "RET", "PIE", "RUF010", "RUF013", "RUF059", "RUF100"]
 "src/schedules/routes.py" = ["PTH", "SIM", "RET", "PIE", "RUF010", "RUF013", "RUF059", "RUF100"]
 "src/collections/routes.py" = ["PTH", "SIM", "RET", "PIE", "RUF010", "RUF013", "RUF059", "RUF100"]
+"src/settings/routes.py" = ["PTH", "SIM", "RET", "PIE", "RUF010", "RUF013", "RUF059", "RUF100"]
 # Same rule for everything lifted out of api_server in Phase 2 Task 8: moved
 # verbatim, so it keeps api_server's ignore set. config_api/routes.py also
 # holds the SSRF-sanitised enable_local_api handler — restructuring its
